+++ melon/rpc/policy/http_rpc_protocol.cc
@@ -35,6 +35,7 @@
 #include "melon/rpc/details/server_private_accessor.h"
 #include "melon/rpc/span.h"
 #include "melon/rpc/socket.h"                       // Socket
+#include "melon/rpc/rpc_dump.h"
 #include "melon/rpc/http_status_code.h"             // HTTP_STATUS_*
 #include "melon/rpc/details/controller_private_accessor.h"
 #include "melon/rpc/builtin/index_service.h"        // IndexService
@@ -1497,6 +1498,16 @@ namespace melon::rpc {
                             return;
                         }
                     }
+                }
+                SampledRequest* sample = AskToBeSampled();
+                if (sample && !is_http2) {
+                    sample->meta.set_compress_type(COMPRESS_TYPE_NONE);
+                    sample->meta.set_protocol_type(PROTOCOL_HTTP);
+                    sample->meta.set_attachment_size(req_body.size());
+
+                    melon::base::end_point ep;
+                    MakeRawHttpRequest(&sample->request, &req_header, ep, &req_body);
+                    sample->submit(start_parse_us);
                 }
             } else {
                 // A http server, just keep content as it is.