+++ test/rpc/rpc_http_rpc_protocol_test.cc
@@ -44,6 +44,15 @@
 #include "melon/strings/starts_with.h"
 #include "melon/strings/ends_with.h"
 #include "melon/fiber/this_fiber.h"
+#include "melon/rpc/rpc_dump.h"
+#include "melon/metrics/collector.h"
+
+namespace melon::rpc {
+    DECLARE_bool(rpc_dump);
+    DECLARE_string(rpc_dump_dir);
+    DECLARE_int32(rpc_dump_max_requests_in_one_file);
+    extern melon::CollectorSpeedLimit g_rpc_dump_sl;
+}
 
 int main(int argc, char *argv[]) {
     testing::InitGoogleTest(&argc, argv);
@@ -1479,6 +1488,90 @@ namespace {
         for (int i = 0; i < req_size; i++) {
             melon::rpc::Join(ids[i]);
         }
+    }
+
+    TEST_F(HttpTest, dump_http_request) {
+        // save origin value of gflag
+        auto rpc_dump_dir = melon::rpc::FLAGS_rpc_dump_dir;
+        auto rpc_dump_max_requests_in_one_file = melon::rpc::FLAGS_rpc_dump_max_requests_in_one_file;
+
+        // set gflag and global variable in order to be sure to dump request
+        melon::rpc::FLAGS_rpc_dump = true;
+        melon::rpc::FLAGS_rpc_dump_dir = "dump_http_request";
+        melon::rpc::FLAGS_rpc_dump_max_requests_in_one_file = 1;
+        melon::rpc::g_rpc_dump_sl.ever_grabbed = true;
+        melon::rpc::g_rpc_dump_sl.sampling_range = melon::COLLECTOR_SAMPLING_BASE;
+
+        // init channel
+        const int port = 8923;
+        melon::rpc::Server server;
+        EXPECT_EQ(0, server.AddService(&_svc, melon::rpc::SERVER_DOESNT_OWN_SERVICE));
+        EXPECT_EQ(0, server.Start(port, nullptr));
+
+        melon::rpc::Channel channel;
+        melon::rpc::ChannelOptions options;
+        options.protocol = "http";
+        ASSERT_EQ(0, channel.Init(melon::base::end_point(melon::base::my_ip(), port), &options));
+
+        // send request and dump it to file
+        {
+            test::EchoRequest req;
+            req.set_message(EXP_REQUEST);
+            std::string req_json;
+            ASSERT_TRUE(json2pb::ProtoMessageToJson(req, &req_json));
+
+            melon::rpc::Controller cntl;
+            cntl.http_request().uri() = "/EchoService/Echo";
+            cntl.http_request().set_content_type("application/json");
+            cntl.http_request().set_method(melon::rpc::HTTP_METHOD_POST);
+            cntl.request_attachment() = req_json;
+            channel.CallMethod(nullptr, &cntl, nullptr, nullptr, nullptr);
+            ASSERT_FALSE(cntl.Failed());
+
+            // sleep 1s, because rpc_dump doesn't run immediately
+            sleep(1);
+        }
+
+        // replay request from dump file
+        {
+            melon::rpc::SampleIterator it(melon::rpc::FLAGS_rpc_dump_dir);
+            melon::rpc::SampledRequest* sample = it.Next();
+            ASSERT_NE(nullptr, sample);
+
+            std::unique_ptr<melon::rpc::SampledRequest> sample_guard(sample);
+
+            // the logic of next code is same as that in rpc_replay.cpp
+            ASSERT_EQ(sample->meta.protocol_type(), melon::rpc::PROTOCOL_HTTP);
+            melon::rpc::Controller cntl;
+            cntl.reset_sampled_request(sample_guard.release());
+            melon::rpc::HttpMessage http_message;
+            http_message.ParseFromCordBuf(sample->request);
+            cntl.http_request().Swap(http_message.header());
+            // clear origin Host in header
+            cntl.http_request().RemoveHeader("Host");
+            cntl.http_request().uri().set_host("");
+            cntl.request_attachment() = http_message.body().movable();
+
+            channel.CallMethod(nullptr, &cntl, nullptr, nullptr, nullptr);
+            ASSERT_FALSE(cntl.Failed());
+            ASSERT_EQ("application/json", cntl.http_response().content_type());
+
+            std::string res_json = cntl.response_attachment().to_string();
+            test::EchoResponse res;
+            json2pb::Json2PbOptions options;
+            ASSERT_TRUE(json2pb::JsonToProtoMessage(res_json, &res, options));
+            ASSERT_EQ(EXP_RESPONSE, res.message());
+        }
+
+        // delete dump directory
+        melon::remove_all(melon::rpc::FLAGS_rpc_dump_dir);
+
+        // restore gflag and global variable
+        melon::rpc::FLAGS_rpc_dump = false;
+        melon::rpc::FLAGS_rpc_dump_dir = rpc_dump_dir;
+        melon::rpc::FLAGS_rpc_dump_max_requests_in_one_file = rpc_dump_max_requests_in_one_file;
+        melon::rpc::g_rpc_dump_sl.ever_grabbed = false;
+        melon::rpc::g_rpc_dump_sl.sampling_range = 0;
     }
 
     TEST_F(HttpTest, spring_protobuf_content_type) {