+++ tools/rpc_replay/rpc_replay.cc
@@ -26,6 +26,7 @@
 #include <melon/rpc/server.h>
 #include <melon/rpc/rpc_dump.h>
 #include <melon/rpc/serialized_request.h>
+#include <melon/rpc/details/http_message.h>
 #include "info_thread.h"
 
 DEFINE_string(dir, "", "The directory of dumped requests");
@@ -40,6 +41,7 @@ DEFINE_string(load_balancer, "", "The algorithm for load balancing");
 DEFINE_int32(timeout_ms, 100, "RPC timeout in milliseconds");
 DEFINE_int32(max_retry, 3, "Maximum retry times");
 DEFINE_int32(dummy_port, 8899, "Port of dummy server(to monitor replaying)");
+DEFINE_string(http_host, "", "Host field for http protocol");
 
 melon::LatencyRecorder g_latency_recorder("rpc_replay");
 melon::counter<int64_t> g_error_count("rpc_replay_error_count");
@@ -157,9 +159,22 @@ static void* replay_thread(void* arg) {
             
             melon::rpc::Controller* cntl = new melon::rpc::Controller;
             req.Clear();
-            
+
+            melon::rpc::SerializedRequest* req_ptr = &req;
+
             cntl->reset_sampled_request(sample_guard.release());
-            if (sample->meta.attachment_size() > 0) {
+
+            if (sample->meta.protocol_type() == melon::rpc::PROTOCOL_HTTP) {
+                melon::rpc::HttpMessage http_message;
+                http_message.ParseFromCordBuf(sample->request);
+                cntl->http_request().Swap(http_message.header());
+                if (!FLAGS_http_host.empty()) {
+                    // reset Host in header
+                    cntl->http_request().SetHeader("Host", FLAGS_http_host);
+                }
+                cntl->request_attachment() = http_message.body().movable();
+                req_ptr = NULL;
+            } else if (sample->meta.attachment_size() > 0) {
                 sample->request.cutn(
                     &req.serialized_data(),
                     sample->request.size() - sample->meta.attachment_size());
@@ -171,13 +186,13 @@ static void* replay_thread(void* arg) {
             const int64_t start_time = melon::get_current_time_micros();
             if (FLAGS_qps <= 0) {
                 chan->CallMethod(nullptr/*use rpc_dump_context in cntl instead*/,
-                        cntl, &req, nullptr/*ignore response*/, nullptr);
+                        cntl, req_ptr, nullptr/*ignore response*/, nullptr);
                 handle_response(cntl, start_time, true);
             } else {
                 google::protobuf::Closure* done =
                     melon::rpc::NewCallback(handle_response, cntl, start_time, false);
                 chan->CallMethod(nullptr/*use rpc_dump_context in cntl instead*/,
-                        cntl, &req, nullptr/*ignore response*/, done);
+                        cntl, req_ptr, nullptr/*ignore response*/, done);
                 const int64_t end_time = melon::get_current_time_micros();
                 int64_t expected_elp = 0;
                 int64_t actual_elp = 0;